operator: Decopule IPAM and node interactions from ENI
