+++ operator/cilium_node.go
@@ -17,6 +17,7 @@ package main
 import (
 	"reflect"
 
+	"github.com/cilium/cilium/pkg/ipam"
 	"github.com/cilium/cilium/pkg/k8s"
 	"github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	"github.com/cilium/cilium/pkg/k8s/informer"
@@ -28,6 +29,21 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 )
+
+var nodeManager *ipam.NodeManager
+
+func ciliumNodeUpdated(resource *v2.CiliumNode) {
+	if nodeManager != nil {
+		// resource is deep copied before it is stored in pkg/aws/eni
+		nodeManager.Update(resource)
+	}
+}
+
+func ciliumNodeDeleted(nodeName string) {
+	if nodeManager != nil {
+		nodeManager.Delete(nodeName)
+	}
+}
 
 func startSynchronizingCiliumNodes() {
 	log.Info("Starting to synchronize CiliumNode custom resources...")
